[refactor] controller의 model 패키지를 dto로 변경
