+++ src/main/java/racingcar/controller/dto/RacingCarInput.java
@@ -1,4 +1,4 @@
-package racingcar.controller.model;
+package racingcar.controller.dto;
 
 import racingcar.service.value.Round;
 import racingcar.utils.Preconditions;
